refactor(actor): Move Actor import to commons for consistency across modules
