+++ shargain/notifications/application/queries/list_notification_configs.py
@@ -2,7 +2,7 @@ import dataclasses
 
 from django.db.models import QuerySet
 
-from shargain.notifications.application.actor import Actor
+from shargain.commons.application.actor import Actor
 from shargain.notifications.application.dto import NotificationConfigDTO
 from shargain.notifications.models import NotificationConfig
