+++ shargain/notifications/tests/application/commands/test_delete_notification_config.py
@@ -2,7 +2,7 @@ import pytest
 
 from shargain.accounts.models import CustomUser
 from shargain.accounts.tests.factories import UserFactory
-from shargain.notifications.application.actor import Actor
+from shargain.commons.application.actor import Actor
 from shargain.notifications.application.commands.delete_notification_config import (
     delete_notification_config,
 )
