+++ shargain/notifications/tests/application/queries/test_get_notification_config.py
@@ -1,7 +1,7 @@
 import pytest
 from django.contrib.auth import get_user_model
 
-from shargain.notifications.application.actor import Actor
+from shargain.commons.application.actor import Actor
 from shargain.notifications.application.dto import NotificationConfigDTO
 from shargain.notifications.application.exceptions import NotificationConfigDoesNotExist
 from shargain.notifications.application.queries.get_notification_config import (
