+++ shargain/notifications/tests/application/queries/test_list_notification_configs.py
@@ -2,7 +2,7 @@ import pytest
 
 from shargain.accounts.models import CustomUser
 from shargain.accounts.tests.factories import UserFactory
-from shargain.notifications.application.actor import Actor
+from shargain.commons.application.actor import Actor
 from shargain.notifications.application.dto import NotificationConfigDTO
 from shargain.notifications.application.queries.list_notification_configs import (
     list_notification_configs,
