+++ shargain/offers/application/commands/delete_target.py
@@ -1,4 +1,4 @@
-from shargain.offers.application.actor import Actor
+from shargain.commons.application.actor import Actor
 from shargain.offers.models import ScrappingTarget
