+++ shargain/offers/application/commands/set_scraping_url_active_status.py
@@ -1,4 +1,4 @@
-from shargain.offers.application.actor import Actor
+from shargain.commons.application.actor import Actor
 from shargain.offers.application.dto import ScrapingUrlDTO
 from shargain.offers.application.exceptions import ScrapingUrlDoesNotExist
 from shargain.offers.models import ScrapingUrl
