+++ shargain/offers/application/commands/toggle_target_active.py
@@ -1,4 +1,4 @@
-from shargain.offers.application.actor import Actor
+from shargain.commons.application.actor import Actor
 from shargain.offers.application.dto import TargetDTO
 from shargain.offers.application.exceptions import TargetDoesNotExist
 from shargain.offers.models import ScrappingTarget
