+++ shargain/offers/application/commands/toggle_target_notifications.py
@@ -4,7 +4,7 @@ Toggle notifications for a ScrapingTarget.
 This command enables or disables notifications for an existing scraping target.
 """
 
-from shargain.offers.application.actor import Actor
+from shargain.commons.application.actor import Actor
 from shargain.offers.application.dto import TargetDTO
 from shargain.offers.application.exceptions import TargetDoesNotExist
 from shargain.offers.models import ScrappingTarget
