+++ shargain/offers/application/commands/update_scraping_url.py
@@ -1,4 +1,4 @@
-from shargain.offers.application.actor import Actor
+from shargain.commons.application.actor import Actor
 from shargain.offers.application.dto import ScrapingUrlDTO
 from shargain.offers.application.exceptions import ScrapingUrlDoesNotExist
 from shargain.offers.models import ScrapingUrl
