+++ shargain/offers/tests/application/commands/test_update_scraping_url.py
@@ -1,6 +1,6 @@
 import pytest
 
-from shargain.offers.application.actor import Actor
+from shargain.commons.application.actor import Actor
 from shargain.offers.application.commands.update_scraping_url import update_scraping_url
 from shargain.offers.application.dto import ScrapingUrlDTO
 from shargain.offers.application.exceptions import ScrapingUrlDoesNotExist
