+++ shargain/offers/tests/application/queries/test_get_scraping_url.py
@@ -1,6 +1,6 @@
 import pytest
 
-from shargain.offers.application.actor import Actor
+from shargain.commons.application.actor import Actor
 from shargain.offers.application.exceptions import ScrapingUrlDoesNotExist
 from shargain.offers.application.queries.get_scraping_url import (
     get_scraping_url,
